requirements.txt
# Upgraded to tensorflow 2.6
flask==2.0.1
gunicorn==20.1.0
tensorflow-cpu==2.6.0
pillow==7.1.2

